Adding level clear and import features; misc bugfixes
- Added 'clear level geometry' feature to SADXLVL2 and SALVL
- Started adding nodetable version 1.5 import support to SAEditorCommon
- fixed obj uv import - uvs were being scaled incorrectly.
- Added data sanitization to Extensions.obj2nj - if a model name is a
literal constant of a C type and would cause compile-time errors,
"model_" is appended to the beginning of the name.
- bounding spheres get calculated at the proper time after levelitem
import
- obj2nj's uv handling now matches the 'new' internal uv storage method
(stored as floats instead of shorts)
- finished adding support for nodetable 1.5
- cleaned up levelitem constructor and implementation abit (position and
rotation are required to construct from file now - this should make it
impossible to forget to supply said information)
- Added File->Import option. This does the same thing as the Add option
in the right-click context menu, except it asks the user if they would
like to clear the level first.
- Added support for *.mtl format extensions
- Paste operation now gives the user a warning that it failed instead of
crashing.