Adding preferences editor to 3d tools, as well as improving the lighting.
- Added EditorOptions class, which houses common editor preferences as
well as providing a single place for the renderers to set their state
(this could be moved to a different class)
- Added EditorOptionsEditor form, which provides a means to edit the
aforementioned options.
- Added a default 3-point light setup for all 3d scenes. This can be
disabled by the user.
- Added a simple defaultLightEditor. It is disabled since users won't
have much of a reason to use it, but in the future we could use it as a
full light editor once the light data is split and documented.